requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
httpx>=0.24.1
pydantic==2.10.4
python-dotenv==1.0.0
supabase>=2.0.0
h2>=4.0.0
anthropic>=0.40.0,<1.0
slowapi>=0.1.9
